pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "agent-redblue-ci"
version = "0.1.0"
description = "Red-team / blue-team security audit for agentic-AI apps — orchestrator and probes."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "Sameer Mohan" }]
keywords = ["security", "ci", "agentic-ai", "prompt-injection", "redteam"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

# Runtime deps kept deliberately small. Scanners themselves (Semgrep, Checkov, etc.)
# are installed separately in the Dockerfile with their own pinned versions.
dependencies = [
    "pyyaml>=6.0.2",
    "jsonschema>=4.23",
    "httpx>=0.27",
    "pydantic>=2.9",
    "click>=8.1",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.39",
    "openai>=1.52",
]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
agent-redblue = "src.orchestrator:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "W"]
ignore = [
    "S101",  # assert in tests
    "S603",  # subprocess call — we wrap tools, this is the point
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
